The Chirp.Razor project resided in another Chirp.Razor folder. The project fodler have been moved, so the path is src/Chirp.Razor/Chirp.Razor.csproj and not src/Chirp.Razor/Chirp.Razor/Chirp.Razor.csproj.
The file DBFacede.cs have been changed, so that there is not two files in the same proeject having top-level statements,
as this is the same as having two static Main method in the same project,
and results in a compile error.
The functions have been refactored into methods of the DB class. Therefore there is no longer use for the GetConnection method,
which has been removed.

The cheep record, have been moved from the DBFacade file to its own CheepRecord file, since best practice is one class per file.

The package reference for SQLite, called Microsoft.Data.SQLite have been added to the Chirp.Razor.csproj file.

Oliver has helped deciding that the ekstra layer of a Chirp.Razor should be moved. He is therefore tagged as a co-author.

Co-authored-by: Oliver <oash@itu.dk>