Added saving to database
Still having trouble with sending the strings between the razor page and the model class. This isn't a priority for now so will leave it as is for now. Also added a IgnoreAntiforgeryTokenAttribute to the builder as it gets an error from the submit otherwise. Will have to read up more on that