Create update on html and everything else
Added all the code Rasmus did on the lecture, and made it work without compile errors or possible null references.
This needs testing, but i'm currently a bit confused about the pages, so we can look at it on tuesday, were we anyway need to probably adjust it some more.