Added user authentication and deleted pages we dont need anymore
Added User.Identity.Authentication.
We can now authenticate a user and get various claims from them, only enabled username for now. The share cheep box only shows when a user is logged in. The userpage only shows when a user is logged in, although it's for now empty. Removed NewCheep and Cheep page and model as they aren't needed anymore aswell as the ChirpBindingModel. Also added a add user in the else statement of adding a cheep so that if the user is logged in and creates a cheep an author is added to the DB and the user can try to add a cheep again and it will work. That implementation need redesigning as this was just a quick implementation for testing it works